add DemoScene, demo video, and README entry for FSM mechanic
Complete the Generic Finite State Machine contribution for issue #15:
- Add DemoScene.unity with Player (7-state FSM), Enemy (6-state FSM),
  ground platform, jumping platforms, patrol waypoints, and UI controls
- Add demo video walkthrough (StateMachineTutorial.mp4)
- Update README with Quick Navigation row and full mechanic card
- Fix folder structure: move scripts to Assets/Scripts/, ScriptExplainer
  to sample root per contributing guide
- Add WhiteSquare sprite asset for character rendering
