+++ README.md
@@ -187,6 +187,7 @@ EventBus.Subscribe<PlayerJumpedEvent>(e => audioManager.PlayJumpSound());
 | 2 | [Generic & Scalable Dialogue System](#2-generic--scalable-dialogue-system) | Mayur | Dialogue | [▶ Watch]
 | 64 | [Utils](#64-Utils) | [Shubham ](https://github.com/vijit101) | Core | [▶ Watch]() |
 (https://github.com/vijit101/UnityMechanicsFramework/tree/main/RuntimeMechanics/Dailogue/2.%20GenericAndScalableDialogueSystem/Assets/Video%20tutorial) |
+| 3 | [Generic Finite State Machine](#3-generic-finite-state-machine) | [AbhinavKRN](https://github.com/AbhinavKRN) | Core | [▶ Watch](Samples~/StateMachineSample/Video/StateMachineTutorial.mp4) |
 
 *More mechanics are added with every merged PR. [Contribute yours →](#9-how-to-contribute)*
 
@@ -317,6 +318,57 @@ dialogueSystem.StartDialogue(npcDatabase, onComplete: () =>
 - Clean separation between data (`DialogueDatabase`) and logic (`DialogueSystem`)
 - Add new conversations without touching any existing scripts
 - Scales to large narrative systems without architectural changes
+
+---
+
+### 3. Generic Finite State Machine
+
+| | |
+|---|---|
+| **Author** | [AbhinavKRN](https://github.com/AbhinavKRN) |
+| **Namespace** | `GameplayMechanicsUMFOSS.Core` |
+| **Location** | `Runtime/Core/StateMachine/StateMachine_UMFOSS.cs` |
+| **Category** | Core |
+| **Demo Scene** | `Samples~/StateMachineSample/Assets/Scenes/DemoScene.unity` |
+| **Video** | [▶ Watch Walkthrough](Samples~/StateMachineSample/Video/StateMachineTutorial.mp4) |
+
+**What it does**
+
+A generic, reusable Finite State Machine that any system in the framework — player controllers, enemy AI, game flow, UI — can use as its backbone without rewriting state logic from scratch. Drop it into any project that needs structured, predictable state transitions with built-in debugging, event publishing, and transition history.
+
+**How to use it**
+
+```csharp
+using GameplayMechanicsUMFOSS.Core;
+
+// Step 1: Create states implementing IState_UMFOSS
+public class IdleState : IState_UMFOSS
+{
+    public void OnEnter() { /* runs once when entering idle */ }
+    public void OnTick() { /* runs every frame while idle */ }
+    public void OnFixedTick() { /* runs every physics step while idle */ }
+    public void OnExit() { /* runs once when leaving idle */ }
+}
+
+// Step 2: Wire up the state machine
+var idle = new IdleState();
+var run = new RunState();
+var fsm = new StateMachine_UMFOSS(idle, gameObject);
+
+// Step 3: Add transitions with conditions
+fsm.AddTransition(idle, run, () => horizontalInput != 0);
+fsm.AddTransition(run, idle, () => horizontalInput == 0);
+fsm.AddAnyTransition(dead, () => health <= 0); // fires from any state
+
+// Step 4: Drive from MonoBehaviour
+void Update() => fsm.Tick();
+void FixedUpdate() => fsm.FixedTick();
+```
+
+**Highlights**
+- Plain C# class (not MonoBehaviour) — testable, reusable, zero coupling to Unity lifecycle
+- Built-in state history, debug inspector overlay, and three EventBus events on every transition
+- Demonstrates the State pattern with anyTransitions for global interrupts (death, stun) that fire from any active state
 
 ---
