+++ Samples~/StateMachineSample/Assets/Scripts/PlayerFSM_UMFOSS.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d0a433ecbc0a42edafec3c1288dab773
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
