+++ Samples~/StateMachineSample/ScriptExplainer.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 50675606e3234847ab2f26e5a5d8965a
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
