+++ Tests/Runtime/StateMachine.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 99e4a2a47f46b415a9c6a65d9f88e6e9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
